# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pyKorf",
    "version": "0.18.0"
  },
  "paths": {
...
    "/api/browse": {
      "get": {
        "tags": [
          "browse"
        ],
        "summary": "Api Browse",
        "description": "Return directory listing for the path browser widget.",
        "operationId": "browseFiles",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Path"
            }
          },
          {
            "name": "filter",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "any",
              "title": "Filter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrowseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/browse/pin": {
      "post": {
        "tags": [
          "browse"
        ],
        "summary": "Pin Folder",
        "description": "Pin a folder for quick access in the path browser.",
        "operationId": "pinFolder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinnedFolderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PinnedFoldersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/browse/unpin": {
      "post": {
        "tags": [
          "browse"
        ],
        "summary": "Unpin Folder",
        "description": "Unpin a folder from quick access.",
        "operationId": "unpinFolder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinnedFolderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PinnedFoldersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/preferences/doc-register": {
      "post": {
        "tags": [
          "preferences"
        ],
        "summary": "Set Doc Register Config",
        "description": "Save Document Register configuration and rebuild DB if Excel path changed.",
        "operationId": "setDocRegisterConfigFromPreferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetDocRegisterConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocRegisterRebuildResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BrowseEntryDir": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "synced": {
            "type": "boolean",
            "title": "Synced",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name",
          "path"
        ],
        "title": "BrowseEntryDir"
      },
      "BrowseEntryFile": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "sharepoint_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sharepoint Url"
          }
        },
        "type": "object",
        "required": [
          "name",
          "path"
        ],
        "title": "BrowseEntryFile"
      },
      "BrowseResponse": {
        "properties": {
          "current": {
            "type": "string",
            "title": "Current"
          },
          "current_sp_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Sp Url"
          },
          "parent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent"
          },
          "drives": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Drives",
            "default": []
          },
          "pinned_folders": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pinned Folders",
            "default": []
          },
          "dirs": {
            "items": {
              "$ref": "#/components/schemas/BrowseEntryDir"
            },
            "type": "array",
            "title": "Dirs",
            "default": []
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/BrowseEntryFile"
            },
            "type": "array",
            "title": "Files",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "current"
        ],
        "title": "BrowseResponse"
      },
...
      "DocRegisterRebuildResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": false
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "stats": {
            "additionalProperties": true,
            "type": "object",
            "title": "Stats",
            "default": {}
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "title": "DocRegisterRebuildResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PinnedFolderRequest": {
        "properties": {
          "folder": {
            "type": "string",
            "title": "Folder"
          }
        },
        "type": "object",
        "required": [
          "folder"
        ],
        "title": "PinnedFolderRequest"
      },
      "PinnedFoldersResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "pinned_folders": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pinned Folders",
            "default": []
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "title": "PinnedFoldersResponse"
      },
...
      "SetDocRegisterConfigRequest": {
        "properties": {
          "excel_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excel Path"
          },
          "sp_site_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sp Site Url"
          }
        },
        "type": "object",
        "title": "SetDocRegisterConfigRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}